renamed Auxiliary to auxiliary
